24/11, 1st upload to new repository for webpage
Created new repository for webpage source code due to a lack of time to completely delete and re-upload all files back to old repository, and creating a new one was quicker. The new version of the webpage features a better file organisation system, a less painfully vivid colour scheme, as suggested by a friend (thanks Euan), and a new section with several pages dedicated to an upcoming project (more info on the site).